+++ tests/unit/commands/using.js
@@ -26,7 +26,8 @@ describe('`using` command', function() {
         };
         this.config = new ConfigMock();
 
-        this.appManager = new AppManager(this.models);
+        this.service = new service.Service(this.config);
+        this.appManager = this.service.appManager;
 
         var app = this.app = this.appManager.buildApp(this.config, {name: 'app'});
         var app2 = this.app2 = this.appManager.buildApp(this.config, {name: 'app2'});