+++ tests/unit/index.js
@@ -31,7 +31,8 @@ describe('CLI', function() {
         this.models = {};
         this.config = new Config();
 
-        this.appManager = new AppManager(this.models);
+        this.service = new service.Service(this.config);
+        this.appManager = this.service.appManager;
         var app = this.app = this.appManager.buildApp(this.config, {name: 'app'});
     });