+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## FUTURE
+
+* [REMOVED] - `bi-config` dependecy
+* [CHANGED] - use `bi-service@>=0.15.x` API
+
 ## v1.2.0
 
 * [CHANGED] - use `bi-service@>0.13.x` API